+++ Model/Notes/GeneratedNote.cs
@@ -51,6 +51,7 @@ public static class GeneratedNoteExtensions
             FromParkingNote(var so) => NoteText.Format(NoteResources.MoveTractionUnitFromParkingToDepartureTrack, so),
             ToParkingNote(var so) => NoteText.Format(NoteResources.MoveTractionUnitToParking, so),
             ReinforcementNote(_, var part) => NoteText.Format(NoteResources.ReinforcesBetweenAnd, part.Train, part.From.OperationLocation, part.To.OperationLocation),
+            TractionUnitExchangeNote(_, var from, var to) => NoteText.Format(NoteResources.TractionUnitExchange, from, to),
             CargoFlowDestinationNote(var part) when part.CargoFlowOptions is not null =>
                 NoteText.Format(NoteResources.BringsWagonsTo, part.ToPlainText),
             _ => string.Empty,